+++ include/menu.h
@@ -3,6 +3,11 @@
 #include <stdio.h>
 
 void open_scores();
+void tuto_menu_text();
+void object_menu_text();
+void object_menu();
+void tuto_menu();
+void title_screen_text();
 void title_screen();
